[accessibility] replace e.g. with "for example"
e.g. being a Latin abbreviation can be confusing for folks for whom English (with all of its weird quirks) is not their first language.  There are better ways of saying "for example" (for example, for example) that will improve readability for all.

see also - https://insidegovuk.blog.gov.uk/2016/07/20/changes-to-the-style-guide-no-more-eg-and-ie-etc/